拆分navi
